use alternate slug to fetch book pages
